# [Numpy] FFI:  atleast_1/2/3d