@Tommliu 2020-03-25
@mxnet-bot run ci [windows-gpu]